requirements.txt
python-frontmatter==1.1.0
Markdown==3.7
Jinja2==3.1.6
